remove background token renew
There are some edge cases where a background token renew wouldn't work
and the client is left without a valid token.
This usually happens when vault is restarted.

These cases (and the usual token renew) can be handled by catching
failed request with StatusCode 403 and retrying them, if a new
authentication attempt succeeds.

This way, all denied requests will be retried with a new valid
token. Therefore, there is no need to renew tokens in the background.